processing libraries when upgrading
