+++ scripts/deploy/truffle-wrapper/artifacts.js
@@ -117,11 +117,28 @@ async function updateArtifact(c, contractAddress, implAddress, libraries) {
     return artifact
 }
 
+async function deployLibrary(name, addresses, signer) {
+    if (addresses[name]) {
+        console.log(`Contract ${name} found from cache`)
+        return addresses[name]
+    } else {
+        const factory = await ethers.getContractFactory(name, signer)
+        const library = await factory.deploy()
+        const h1 = library.deployTransaction.hash
+        await library.deployed()
+        const address = (await web3.eth.getTransactionReceipt(h1)).contractAddress
+        console.log(`Library ${name} deployed into address ${address}`)
+        addresses[name] = address
+        return address
+    }
+}
+
 module.exports = {
     updateArtifact,
     writeArtifact,
     readArtifact,
     exportLibraryArtifacts,
     exportLibraryArtifact,
-    exportArtifacts
+    exportArtifacts,
+    deployLibrary
 }
